Move process functions to pkg/system
